Files update

+++ 2_Files_manipulation/files.py
@@ -12,15 +12,24 @@ class fileManipulator():
         # Path to current workin directory in p variable
         p = Path.cwd()
 
-        try:
-            shutil.copy(p / src, p / dest)
-        # If dest contains folders that doesn't exist
-        # then create them
-        except FileNotFoundError:
-            dir_to_create = dest[:dest.rfind('/')]
-            os.makedirs(dir_to_create)
-            shutil.copy(p / src, p / dest)
-            logging.info(f"created '{dir_to_create}' folders")
+        if Path(p / src).exists():
+            try:
+                if Path(p / src).is_dir():
+                    shutil.copytree(p / src, p /dest)
+                else:
+                    shutil.copy(p / src, p / dest)
+            # If dest contains folders that doesn't exist
+            # then create them
+            except FileNotFoundError:
+                dir_to_create = dest[:dest.rfind('/')]
+                os.makedirs(dir_to_create)
+                if Path(p / src).is_dir():
+                    shutil.copytree(p / src, p /dest)
+                else:
+                    shutil.copy(p / src, p / dest)
+                logging.info(f"created '{dir_to_create}' folders")
+        else:
+            logging.error(f"{src} doesn't exist")
 
     # Moves src to dest
     def move(self, src, dest):
@@ -29,15 +38,18 @@ class fileManipulator():
         # Path to current workin directory in p variable
         p = Path.cwd()
 
-        try:
-            shutil.move(p / src, p / dest)
-        # If dest contains folders that doesn't exist
-        # then create them
-        except FileNotFoundError:
-            dir_to_create = dest[:dest.rfind('/')]
-            os.makedirs(dir_to_create)
-            shutil.move(p / src, p / dest)
-            logging.info(f"created '{dir_to_create}' folders")
+        if Path(p / src).exists():
+            try:
+                shutil.move(p / src, p / dest)
+            # If dest contains folders that doesn't exist
+            # then create them
+            except FileNotFoundError:
+                dir_to_create = dest[:dest.rfind('/')]
+                os.makedirs(dir_to_create)
+                shutil.move(p / src, p / dest)
+                logging.info(f"created '{dir_to_create}' folders")
+        else:
+            logging.error(f"{src} doesn't exits")
 
     # Creates zip-archive named [name] with [dir] in it
     def zip(self, dir, name):
@@ -70,6 +82,7 @@ if __name__ == '__main__':
     logging.basicConfig(filename='files.log', filemode='w', level=logging.INFO, format='%(asctime)s | %(levelname)s | %(funcName)s, %(lineno)d: %(message)s')
     fm = fileManipulator()
     fm.copy('files.log', 'logss/files.log')
-    fm.move('logss/files.log', 'logs/files.log')
+    fm.copy('logss', 'logsss/logss')
+    fm.move('test', 'logss')
     fm.zip('test', 'test.zip')
     fm.unzip('test.zip')
